+++ flags/models.py
@@ -12,12 +12,21 @@ class FlagState(models.Model):
         unique_together = ("name", "condition", "value")
 
     def __str__(self):
+        required_str = " (required)" if self.required else ""
         return (
-            "{name} is enabled when {condition} is "
-            "{value}{required}".format(
-                name=self.name,
-                condition=self.condition,
-                value=self.value,
-                required=" (required)" if self.required else "",
-            )
+            f"{self.name} is enabled when {self.condition} is "
+            f"{self.value}{required_str}"
         )
+
+
+class FlagMetadata(models.Model):
+    name = models.CharField(max_length=64)
+    key = models.CharField(max_length=64)
+    value = models.TextField()
+
+    class Meta:
+        app_label = "flags"
+        unique_together = ("name", "key")
+
+    def __str__(self):
+        return f"{self.name} ({self.key}: {self.value})"