+++ flags/tests/test_models.py
@@ -1,6 +1,6 @@
 from django.test import TestCase
 
-from flags.models import FlagState
+from flags.models import FlagMetadata, FlagState
 
 
 class FlagStateTestCase(TestCase):
@@ -16,4 +16,14 @@ class FlagStateTestCase(TestCase):
         )
         self.assertEqual(
             str(state), "MY_FLAG is enabled when boolean is True (required)"
+        )
+
+
+class FlagMetadataTestCase(TestCase):
+    def test_flag_str(self):
+        metadata = FlagMetadata.objects.create(
+            name="MY_FLAG", key="help_text", value="enable a cool thing"
+        )
+        self.assertEqual(
+            str(metadata), "MY_FLAG (help_text: enable a cool thing)"
         )